Move Language settings to `menus_settings_language.cpp`
